// resources/js/api/generated/models/userInfo.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { RoleInfo } from './roleInfo';

export interface UserInfo {
  created_at: string;
  email: string;
  /** @nullable */
  email_verified_at?: string | null;
  /** @minimum 0 */
  id: number;
  /** @nullable */
  last_login_at?: string | null;
  roles?: RoleInfo[];
  totp_enabled: boolean;
  updated_at: string;
  username: string;
}